Moved input types to its own file
